+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - portscan: Report commented `PORTEPOCH` or `PORTREVISION` lines
   via new `lint.commented-portrevision` lint, selectable with
   `-o comments`, enabled by default
+- portedit: `apply list` will now print a list of available edits
 
 ### Changed